fix(Navbar): refactor NavDropdown
